requirements.txt
streamlit
ollama
bcrypt
cryptography
mysql-connector-python
pypdf2
python-docx